Ensure parameterized tests don't produce overlapping snapshot files.
The current escaping function replaces non-alphanumeric characters with
the "_" character to ensure they are compatible with the filesystem's
constrainted. However, this normalization can cause a problem when two
separate parameterized tests point at the same snapshot file after this
normalization step.

This change modifies the normalization function to replace these
characters with either their Unicode name or hexidecimal escape code.
The resulting filenames are a bit verbose, but this approach ensures
that these snapshots are always creating unique files.

If you revert the normalization function change, you'll note that the
"overlap]" and "overlap[" tests in `test_fmt.py` will refer to the
same snapshot file and the tests will fail and/or be flaky.